feat: adding util paste with generals functions important for the projerct
