+++ agendamento-chacara/src/utils/remove-non-numerics.ts
@@ -1,0 +1,4 @@
+export const removeNonNumerics = (str: string) => {
+    return str.replace(/[^\d]/g, "");
+  };
+
